requirements.txt
python-telegram-bot[job-queue]==21.8
httpx==0.27.2
requests==2.32.3
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
openai==1.43.0

